Split the current snapshot interfaces to two
The managers used to inherit from the snapshot interface
to make one off queries easier but this a bit too fragile
as it allows passing the manager itself as a snapshot which
is not true.

Instead moving these shared methods to their own interface
to allow both the manager and snapshot to implement them
without breaking this contract and prevent potential bugs.

There is still small risk this could be used wrong but
imo this is better for usability.